libraries/serializers.py
from rest_framework import serializers
from libraries.models import Accounts


class AccountsSerializer(serializers.ModelSerializer):
    class Meta:
        model = Accounts
        fields = '__all__'
