requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
defusedxml==0.7.1
diff-match-patch==20230430
dj-rest-auth==4.0.1
Django==4.2.4
django-cors-headers==4.2.0
django-environ==0.10.0
django-import-export==3.2.0
djangorestframework==3.14.0
docopt==0.6.2
drf-yasg==1.21.7
et-xmlfile==1.1.0
idna==3.4
inflection==0.5.1
MarkupPy==1.14
mysqlclient==2.2.0
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
pipreqs==0.4.13
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
xlrd==2.0.1
xlwt==1.3.0
yarg==0.1.9
